% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practicalgg.R
\docType{package}
\name{practicalgg}
\alias{practicalgg}
\alias{practicalgg-package}
\title{Practical ggplot2, a package demonstrating practical uses of the ggplot2 package.}
\description{
Practical ggplot2, a package demonstrating practical uses of the ggplot2 package.
}
