% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy.R
\docType{data}
\name{happy}
\alias{happy}
\title{Data related to happiness from the general social survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51020 rows and 10 columns.}
\usage{
happy
}
\description{
This dataset provides a small sample of variables related to happiness from the general social survey (GSS).
It is a reexport of the \code{\link[productplots:happy]{productplots::happy}} dataset, converted into a tibble.
}
\examples{
library(tidyverse)

happy \%>\%
  na.omit() \%>\%
  mutate(health = fct_rev(health)) \%>\%
  ggplot(aes(x = age, y = stat(count))) +
  geom_density(fill = "lightblue") +
  facet_wrap(~health)
}
\keyword{datasets}
